Move ssm_retry
